requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-core>=0.1.0
databricks-sdk>=0.18.0
python-dotenv>=1.0.0
openai>=1.12.0
requests>=2.31.0
tenacity>=8.2.3

